+++ pkg/util/test/fake_snapshot_service.go
@@ -52,7 +52,7 @@ func (s *FakeSnapshotService) CreateSnapshot(volumeID string) (string, error) {
 	return s.SnapshottableVolumes[volumeID].SnapshotID, nil
 }
 
-func (s *FakeSnapshotService) CreateVolumeFromSnapshot(snapshotID, volumeType string, iops *int) (string, error) {
+func (s *FakeSnapshotService) CreateVolumeFromSnapshot(snapshotID, volumeType string, iops *int64) (string, error) {
 	key := api.VolumeBackupInfo{
 		SnapshotID: snapshotID,
 		Type:       volumeType,
@@ -72,7 +72,7 @@ func (s *FakeSnapshotService) DeleteSnapshot(snapshotID string) error {
 	return nil
 }
 
-func (s *FakeSnapshotService) GetVolumeInfo(volumeID string) (string, *int, error) {
+func (s *FakeSnapshotService) GetVolumeInfo(volumeID string) (string, *int64, error) {
 	if volumeInfo, exists := s.SnapshottableVolumes[volumeID]; !exists {
 		return "", nil, errors.New("VolumeID not found")
 	} else {
